feat: added home feed screen for chat in core
